지금 깃 허브가 작동하지 않아서 터미널 열어서 대충 푸쉬후 나중에 변경된 내역 다시 업로드함.

+++ Assets/Script/BaseCharacter.cs
@@ -22,8 +22,8 @@ public class BaseCharacter : MonoBehaviour
     public float curMP;
     public float maxMP;
 
-    public float baseAttackPower;
-    public float curAttackPower;
+    public int   baseAttackPower;
+    public int   curAttackPower;
 
     public float baseAttackSpeed;
     public float curAttackSpeed;
